Create Mark-and-Toys-vector-sorting.cpp
Mark and Jane are very happy after having their first kid. Their son loves toys, so Mark wants to buy some. There are a number of different toys lying in front of him, tagged with their prices. Mark has only a certain amount to spend, and he wants to maximize the number of toys he buys with this money.