Create Sherlock-valid-string.py
Sherlock considers a string to be valid if all characters of the string appear the same number of times. It is also valid if he can remove just  1 character at 1  index in the string, and the remaining characters will occur the same number of times. Given a string , determine if it is valid.